dbs-boot: fix arm compile error
move Error from fdt to mod to fix arm compile error
